% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stat_generator.R
\name{multi_sample_stats}
\alias{multi_sample_stats}
\title{Generate multiple series of statistics for a supplied series of sample data.}
\usage{
multi_sample_stats(.data, stat_list = list(mean), ...)
}
\arguments{
\item{.data}{A data frame or data frame housed within a list}

\item{stat_list}{The statistics to be run on each element of the data frame. This
needs to be a user-defined or package function that runs a statistic. This
function is intended to take in a list of functions, so the target parameter
is intended to be a list object. The default is a single element list with
\code{mean} as the function}

\item{...}{Ellipsis intended to pass along additional arguments for the supplied
statistic}
}
\value{
A list of equal length vectors
}
\description{
Generate multiple series of statistics for a supplied series of sample data.
}
\examples{
BLANK
}
