% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_generator.R
\name{sample_data}
\alias{sample_data}
\title{Generate a series of equal sized samples from a user-defined distribution}
\usage{
sample_data(num_samples = 25, sample_size = 50, dist_size = 100, max_val = 100)
}
\arguments{
\item{num_samples}{The total number of samples to create}

\item{sample_size}{The total number of measurements in each sample. Must be one less than dist_size.}

\item{dist_size}{The total number of measurements within a hypothetical population distribution.}

\item{max_val}{The maximum number a value can assume within the population.}
}
\value{
a data frame
}
\description{
Generate a series of equal sized samples from a user-defined distribution
}
\examples{
head(sample_data())
}
