% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_generator.R
\name{sample_stat}
\alias{sample_stat}
\title{Generate a series of statistics for a supplied series of sample data.}
\usage{
sample_stat(.data, stat = mean, ...)
}
\arguments{
\item{.data}{A data frame or data frame housed within a list}

\item{stat}{The statistic to be run on each element of the data frame. This
needs to be a user-defined or package function that runs a statistic. The default
is \code{mean}}

\item{...}{Ellipsis intended to pass along additional arguments for the supplied
statistic}
}
\value{
A numeric vector
}
\description{
Generate a series of statistics for a supplied series of sample data.
}
\examples{
BLANK
}
