% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_generator.R
\name{sample_stats}
\alias{sample_stats}
\title{Generate a series of statistics for a supplied series of sample data.}
\usage{
sample_stats(samp_data, stat = mean)
}
\arguments{
\item{samp_data}{A data frame of sample data}

\item{stat}{The statistic to be run on each element of the data frame}
}
\value{
A numeric vector
}
\description{
Generate a series of statistics for a supplied series of sample data.
}
